# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Sven Gateway API
  version: 1.0.0
  description: |
    REST + SSE API for the Sven companion app gateway.

    **Authentication:** All endpoints except `/v1/auth/login`,
    `/v1/auth/refresh`, and `/v1/config/deployment` require a Bearer token
    obtained from `/v1/auth/login` or `/v1/auth/refresh`.

    **Envelope format:** Successful responses use a standard envelope:
    ```json
    { "data": { ... } }        // single object
    { "data": { "rows": [...], "has_more": true } }  // paginated list
    ```
    Error responses use:
    ```json
    { "error": { "code": "ERR_CODE", "message": "Human-readable message" } }
    ```

    **Base URL:** Configured per environment via `SVEN_API_BASE` dart-define.
    Production: `https://app.example.com`
    Development: `https://dev.example.com`
  contact:
    name: 47network / openclaw-sven
  license:
    name: Proprietary
servers:
  - url: https://app.example.com
    description: Production
  - url: https://dev.example.com
    description: Development

# ─── Security ──────────────────────────────────────────────────────────────────
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ─── Reusable Schemas ────────────────────────────────────────────────────────
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: UNAUTHORIZED
            message:
              type: string
              example: "Authentication token is missing or invalid"

    # ── Auth ──
...
    DeploymentSetupRequest:
      type: object
      required: [mode, username, password]
      properties:
        mode:
          type: string
          enum: [personal, enterprise, demo]
        username:
          type: string
        password:
          type: string
          format: password
          minLength: 8
        display_name:
          type: string
          nullable: true

    # ── Devices ──
...
  responses:
...
    UnprocessableEntity:
      description: Validation error — request body is malformed or missing required fields
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

# ─── Paths ─────────────────────────────────────────────────────────────────────
paths:
...
  /v1/config/deployment/setup:
    post:
      operationId: setupDeployment
      tags: [Config]
      summary: One-time deployment initialisation
      description: |
        Called once during first-run setup to create the initial admin user and
        configure the deployment mode. The endpoint is disabled after first use.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeploymentSetupRequest'
      responses:
        '200':
          description: Deployment initialised; admin user created
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    description: Newly created admin user info
        '409':
          description: Deployment has already been initialised
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

  # ═══════════════════════════════════════════════════════════════════════════
  # Admin — Device Management
  # ═══════════════════════════════════════════════════════════════════════════
...